pyproject.toml
[project]
name = "agent-test-harness"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyaml>=24.9.0",
    "pycobertura>=3.3.2",
    "requests>=2.32.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["agent_test_harness"]

[project.scripts]
agent-test-harness = "agent_test_harness.main:main"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=agent_test_harness --cov-report=term-missing"
